requirements.txt
fastapi
uvicorn
requests
groq
supabase
argon2-cffi
python-dotenv
openai